% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-cite-keys.R
\name{extract_cite_keys}
\alias{extract_cite_keys}
\title{Extract cite keys}
\usage{
extract_cite_keys(doc)
}
\arguments{
\item{doc}{a string character}
}
\value{
a character vector of extracted cite keys
}
\description{
Extract cite keys like "@foo:2014" from a text (.Rmd, .md) file.
}
\examples{
extract_cite_keys("@smith:2015")
extract_cite_keys("@smith2015first")
}
