% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce-bib.R
\name{read_bib}
\alias{read_bib}
\title{Read a .bib file}
\usage{
read_bib(file, quiet = TRUE, ...)
}
\arguments{
\item{file}{Path to the .bib file.}

\item{quiet}{Show \code{\link[RefManageR:ReadBib]{RefManageR::ReadBib()}} output/warnings?}

\item{...}{Arguments passed to \code{\link[RefManageR:ReadBib]{RefManageR::ReadBib()}}}
}
\description{
Thin wrapper around \code{\link[RefManageR:ReadBib]{RefManageR::ReadBib()}} that suppresses warnings and
messages.
}
\examples{
\dontrun{
  master <- read_bib("master.bib")
}

}
