% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce-bib.R
\name{reduce_bib}
\alias{reduce_bib}
\title{Reduce a central .bib file down to used elements}
\usage{
reduce_bib(file, master_bib, out_bib, ...)
}
\arguments{
\item{file}{path to the target .Rmd file}

\item{master_bib}{path to whistle/master.bib (or any other central .bib file
for that matter)}

\item{out_bib}{path/name of the output .bib file to write}

\item{...}{Arguments passed to \code{\link[RefManageR:ReadBib]{RefManageR::ReadBib()}}}
}
\description{
Condense a master.bib file to only the references that are used in a paper.
}
