% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce-bib.R
\name{strip_cite_keys}
\alias{strip_cite_keys}
\title{Strip cite keys}
\usage{
strip_cite_keys(doc)
}
\arguments{
\item{doc}{a string character}
}
\value{
a character vector of extracted cite keys
}
\description{
Strip cite keys like "@foo:2014" from a text (.Rmd, .md) file.
}
\examples{
strip_cite_keys("@smith:2015")
strip_cite_keys("@smith2015first")
}
