% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce-bib.R
\name{write_bib}
\alias{write_bib}
\title{Write a .bib file}
\usage{
write_bib(bib, ...)
}
\arguments{
\item{bib}{A \code{\link[RefManageR:BibEntry]{RefManageR::BibEntry()}} object.}

\item{...}{Arguments passed to \code{\link[RefManageR:WriteBib]{RefManageR::WriteBib()}}}
}
\description{
Thin wrapper around \code{\link[RefManageR:WriteBib]{RefManageR::WriteBib()}}.
}
